feat(react): fix template issues
