+++ source/app/components/Dashboard.tsx
@@ -1,0 +1,14 @@
+// packages
+import React from 'react';
+import Button from '@mui/material/Button';
+
+const Dashboard = () => {
+  return (
+    <>
+      <h1>Home</h1>
+      <Button variant="contained">Hello World</Button>
+    </>
+  );
+};
+
+export default Dashboard;
