--- source/app/components/Home.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export const Home = () => {
-  return <h1>Home</h1>;
-};
