+++ source/app/components/Settings.tsx
@@ -1,0 +1,12 @@
+// packages
+import React from 'react';
+
+const Settings = () => {
+  return (
+    <div>
+      <h1>About</h1>
+    </div>
+  );
+};
+
+export default Settings;
