+++ source/package.json
@@ -42,31 +42,33 @@
     "watch:view": "webpack --watch --mode development"
   },
   "devDependencies": {
-    "@types/glob": "7.2.0",
-    "@types/mocha": "9.1.1",
-    "@types/node": "18.0.0",
-    "@types/react": "18.0.13",
-    "@types/react-dom": "18.0.5",
-    "@types/vscode": "1.68.0",
-    "@typescript-eslint/eslint-plugin": "5.28.0",
-    "@typescript-eslint/parser": "5.28.0",
-    "css-loader": "6.7.1",
-    "eslint": "8.17.0",
-    "eslint-plugin-react-hooks": "4.6.0",
-    "glob": "8.0.3",
-    "mocha": "10.0.0",
-    "npm-run-all": "4.1.5",
-    "prettier": "2.7.1",
-    "style-loader": "3.3.1",
-    "ts-loader": "9.3.0",
-    "typescript": "4.7.3",
-    "vscode-test": "1.4.1",
-    "webpack": "5.73.0",
-    "webpack-cli": "4.10.0"
+    "@types/glob": "^7.2.0",
+    "@types/mocha": "^9.1.1",
+    "@types/node": "^18.0.0",
+    "@types/react": "^18.0.14",
+    "@types/vscode": "^1.68.0",
+    "@typescript-eslint/eslint-plugin": "^5.28.0",
+    "@typescript-eslint/parser": "^5.28.0",
+    "css-loader": "^6.7.1",
+    "eslint": "^8.17.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "glob": "^8.0.3",
+    "mocha": "^10.0.0",
+    "npm-run-all": "^4.1.5",
+    "prettier": "^2.7.1",
+    "style-loader": "^3.3.1",
+    "ts-loader": "^9.3.0",
+    "typescript": "^4.7.3",
+    "vscode-test": "^1.4.1",
+    "webpack": "^5.73.0",
+    "webpack-cli": "^4.10.0"
   },
   "dependencies": {
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-router-dom": "6.3.0"
+    "@emotion/react": "^11.9.3",
+    "@emotion/styled": "^11.9.3",
+    "@mui/material": "^5.8.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.3.0"
   }
 }
